refactor: split daily command family
